requirements.txt
requests>=2.32.3
pyyaml>=5.4.1